optimize: drastically reduce memory footprint by using QByteArray instead of QString
we only convert to QString on demand when we actually display a string
to the user. most often, this is just a small subset of the data
we have lying around in total.

the memory consumption gain is esp. noticeable for big data files,
where it reduces the memory consumption by ~40%