Optimize: force the use of the const overload of QVector::first.
Value semantics can be tricky, sometimes. Returning a QVector<...>
by value is fine, but across library boundaries the NRVO is not
kicking in, and we end up with a refcount increment. Then, after that,
the compiler picks the non-const overload of QVector::first, which
then triggers a detach as we just incremented the refcount.