Refactor the MDI implementation.
The goal is to uncouple the code, esp. the mainwindow from the
single documents. Furthermore, this makes it simpler to add more
per-document views in the future.

We make the documents KXMLGUIClients and also add separate widgets
for the single tabs, i.e. memory evolution and call-graph currently.
These care for the dynamic addition of the actions and their state.

The selection code is also cleaned up to better follow QAIM best-
practices (i.e. using idx.data() with custom roles). It's still
a bit nasty since it has to sync with the global data tree model
shown in the dock widget. Something that can be cleaned up in the
future.