Port and reenable KGraphViewer support
Summary:
I never used the Callgraph tab in qt4 versions, so cannot tell if there
are regressions. There surely is room for improvements,
but exposing the current state might trigger people to work on it.

Needs latest frameworks branch of KGraphViewer, which currently
is prepared to be merged to master and released as v2.4 in a few weeks.

Reviewers: mwolff, ltoscano

Reviewed By: mwolff

Differential Revision: https://phabricator.kde.org/D5244