Actually install own icon file
It's a copy of oxygen's which I guess is the one that was meant to be
used. It needs to be specified otherwise the icon depends on the system
icon theme, which looks especially weird in software centers.
Also it won't be accepted by flatpak.

CCMAIL: milian.wolff@kdab.com